test: move the identity-lifecycle tests into their own file
They were appended to the end of cmd_test.go, which collides with any other
branch that does the same. In their own file they merge cleanly regardless of
the order these land in.
